Skipping some tests against older server
